+++ cmd/tempo/app/overrides_validation.go
@@ -8,6 +8,7 @@ import (
 	"github.com/grafana/tempo/modules/generator"
 	"github.com/grafana/tempo/modules/overrides/userconfigurable/api"
 	"github.com/grafana/tempo/modules/overrides/userconfigurable/client"
+	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 )
 
 type overridesValidator struct {
@@ -44,6 +45,14 @@ func (v *overridesValidator) Validate(limits *client.Limits) error {
 		for p := range processors.GetMap() {
 			if !slices.Contains(generator.SupportedProcessors, p) {
 				return fmt.Errorf("metrics_generator.processor \"%s\" is not a known processor, valid values: %v", p, generator.SupportedProcessors)
+			}
+		}
+	}
+
+	if filterPolicies, ok := limits.GetMetricsGenerator().GetProcessor().GetSpanMetrics().GetFilterPolicies(); ok {
+		for _, fp := range filterPolicies {
+			if err := filterconfig.ValidateFilterPolicy(fp); err != nil {
+				return err
 			}
 		}
 	}